pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "numpy==1.24.3",
    "pillow>=11.1.0",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "tensorflow>=2.14.0",
    "werkzeug>=3.1.3",
]
